Inline cumulus motif-tagging controller in base application
