--- collections-online-cumulus/plugins/motif-tag-controller.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  type: 'motif-tag-controller',
-  module: require('../controllers/motif-tagging'),
-  initialize: () => {}
-};
